+++ src/Templates/Template.php
@@ -40,7 +40,7 @@ abstract class Template implements TemplateInterface
     /**
      * string VendorVersion.
      */
-    protected $ver = '<b>Ver:</b> dev.2402.0903';
+    protected $ver = '<b>Ver:</b> dev.2402.2116';
 
     /**
      * TemplateDefault constructor.
@@ -168,7 +168,7 @@ abstract class Template implements TemplateInterface
 
     /**
      * @param  string  $key
-     * @return string
+     * @return string|void
      *
      * @throws \DaveJamesMiller\Breadcrumbs\Exceptions\InvalidBreadcrumbException
      * @throws \DaveJamesMiller\Breadcrumbs\Exceptions\UnnamedRouteException